move utils to plugin folder
