+++ spyder/plugins/editor/utils/switcher.py
@@ -15,7 +15,7 @@ import os.path as osp
 from spyder.config.base import _
 from spyder.config.manager import CONF
 from spyder.utils.icon_manager import ima
-from spyder.utils.switcher import shorten_paths, get_file_icon
+from spyder.plugins.switcher.utils import shorten_paths, get_file_icon
 from spyder.plugins.completion.api import SymbolKind, SYMBOL_KIND_ICON
