+++ spyder/plugins/switcher/utils.py
