+++ spyder/plugins/switcher/widgets/switcher.py
@@ -28,7 +28,7 @@ from spyder.py3compat import TEXT_TYPES, to_text_string
 from spyder.utils.icon_manager import ima
 from spyder.widgets.helperwidgets import HTMLDelegate
 from spyder.utils.stringmatching import get_search_scores
-from spyder.utils.switcher import clean_string
+from spyder.plugins.switcher.utils import clean_string
 
 # Style dict constants
 FONT_SIZE = 10
